requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bitarray==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
cytoolz==1.0.1
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
hexbytes==1.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.2
parsimonious==0.10.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pyunormalize==16.0.0
pywin32==308
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.7.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
